fix smtp timeout issue.
the previous shadowsocks implementation assumes that for all protocols
based on tcp, the first data is from client to server. so it send
ss header to remote server only when client sends the first data.
the assumption is not true for smtp protocol, as the first data in smtp
is from server to client. so sending mail with the system mail app will
fail with timeout.

the fix is to send the ss header to remote server once tcp connection
established, so that the remote server know which service to connect and
make a connection.